requirements.txt
asgiref==3.5.0
Django==4.0.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2021.3
tzdata==2021.5
djongo==1.3.6
mongoengine==0.24.0
pymongo==3.12.1

